refactor: refactor time capsule item component
